wip: refactor the training script into multiple files
