+++ scripts/evaluate_models.py
@@ -1,0 +1,20 @@
+
+from dataclasses import dataclass
+import torch
+import numpy as np
+import tifffile
+from pathlib import Path
+import os
+import argparse
+import matplotlib.pyplot as plt
+
+
+
+def evaluate_model() -> None:
+    pass
+
+def main():
+    pass
+
+if __name__ == "__main__":
+    main()
